Fix a bug in the packager.
Fix a bug where the packager would crash if no previously package was found.

+++ src/packager.py
@@ -8,7 +8,8 @@ def build_package():
     infilename = "config/packagefiles.txt"
     infile = open(infilename)
     
-    os.remove("clientpackage.tar.gz")
+    if os.path.exists("clientpackage.tar.gz"):
+        os.remove("clientpackage.tar.gz")
     
     tar = tarfile.open("clientpackage.tar.gz", "w:gz")
